+++ FinancesBackend/src/main/java/com/My/FinancesApp/dto/IncomeDto.java
@@ -3,12 +3,15 @@ package com.My.FinancesApp.dto;
 import lombok.Data;
 import com.My.FinancesApp.model.CategoryIncome;
 
+import java.math.BigDecimal;
+
 
 @Data
 public class IncomeDto {
 
+
     private String name;
     private String description;
-    private double value;
+    private BigDecimal value;
     private CategoryIncome category;
 }
